+++ authentication/typescript/azure/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./azureIdentityAuthenticationProvider";
+export * from './azureIdentityAuthenticationProvider';
+export * from './azureIdentityAccessTokenProvider';