+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PHP Request Adapter (includes middleware) #1048, #918, #1024, #1025
 - Added support for PHP Json Serialization.
 - Update PHP abstractions to make property keys and values nullable in `SerializationWriter.php`.
+- Added Python abstractions library.
+- Adds hostname and protocol validation in authentication. #1051
 
 ### Changed